+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/metastore/cache/CachedStore.java
@@ -1701,133 +1701,6 @@ public class CachedStore implements RawStore, Configurable {
     return partitions;
   }
 
-  @Override public boolean addRole(String rowName, String ownerName)
-      throws InvalidObjectException, MetaException, NoSuchObjectException {
-    return rawStore.addRole(rowName, ownerName);
-  }
-
-  @Override public boolean removeRole(String roleName) throws MetaException, NoSuchObjectException {
-    return rawStore.removeRole(roleName);
-  }
-
-  @Override public boolean grantRole(Role role, String userName, PrincipalType principalType, String grantor,
-      PrincipalType grantorType, boolean grantOption)
-      throws MetaException, NoSuchObjectException, InvalidObjectException {
-    return rawStore.grantRole(role, userName, principalType, grantor, grantorType, grantOption);
-  }
-
-  @Override public boolean revokeRole(Role role, String userName, PrincipalType principalType, boolean grantOption)
-      throws MetaException, NoSuchObjectException {
-    return rawStore.revokeRole(role, userName, principalType, grantOption);
-  }
-
-  @Override public PrincipalPrivilegeSet getUserPrivilegeSet(String userName, List<String> groupNames)
-      throws InvalidObjectException, MetaException {
-    return rawStore.getUserPrivilegeSet(userName, groupNames);
-  }
-
-  @Override public PrincipalPrivilegeSet getDBPrivilegeSet(String catName, String dbName, String userName,
-      List<String> groupNames) throws InvalidObjectException, MetaException {
-    return rawStore.getDBPrivilegeSet(catName, dbName, userName, groupNames);
-  }
-
-  @Override public PrincipalPrivilegeSet getConnectorPrivilegeSet(String catName, String connectorName, String userName,
-      List<String> groupNames) throws InvalidObjectException, MetaException {
-    return rawStore.getConnectorPrivilegeSet(catName, connectorName, userName, groupNames);
-  }
-
-  @Override public PrincipalPrivilegeSet getTablePrivilegeSet(String catName, String dbName, String tableName,
-      String userName, List<String> groupNames) throws InvalidObjectException, MetaException {
-    return rawStore.getTablePrivilegeSet(catName, dbName, tableName, userName, groupNames);
-  }
-
-  @Override public PrincipalPrivilegeSet getPartitionPrivilegeSet(String catName, String dbName, String tableName,
-      String partition, String userName, List<String> groupNames) throws InvalidObjectException, MetaException {
-    return rawStore.getPartitionPrivilegeSet(catName, dbName, tableName, partition, userName, groupNames);
-  }
-
-  @Override public PrincipalPrivilegeSet getColumnPrivilegeSet(String catName, String dbName, String tableName,
-      String partitionName, String columnName, String userName, List<String> groupNames)
-      throws InvalidObjectException, MetaException {
-    return rawStore.getColumnPrivilegeSet(catName, dbName, tableName, partitionName, columnName, userName, groupNames);
-  }
-
-  @Override public List<HiveObjectPrivilege> listPrincipalGlobalGrants(String principalName,
-      PrincipalType principalType) {
-    return rawStore.listPrincipalGlobalGrants(principalName, principalType);
-  }
-
-  @Override public List<HiveObjectPrivilege> listPrincipalDBGrants(String principalName, PrincipalType principalType,
-      String catName, String dbName) {
-    return rawStore.listPrincipalDBGrants(principalName, principalType, catName, dbName);
-  }
-
-  @Override public List<HiveObjectPrivilege> listPrincipalDCGrants(String principalName, PrincipalType principalType,
-                                                                   String dcName) {
-    return rawStore.listPrincipalDCGrants(principalName, principalType, dcName);
-  }
-
-  @Override public List<HiveObjectPrivilege> listAllTableGrants(String principalName, PrincipalType principalType,
-      String catName, String dbName, String tableName) {
-    return rawStore.listAllTableGrants(principalName, principalType, catName, dbName, tableName);
-  }
-
-  @Override public List<HiveObjectPrivilege> listPrincipalPartitionGrants(String principalName,
-      PrincipalType principalType, String catName, String dbName, String tableName, List<String> partValues,
-      String partName) {
-    return rawStore
-        .listPrincipalPartitionGrants(principalName, principalType, catName, dbName, tableName, partValues, partName);
-  }
-
-  @Override public List<HiveObjectPrivilege> listPrincipalTableColumnGrants(String principalName,
-      PrincipalType principalType, String catName, String dbName, String tableName, String columnName) {
-    return rawStore
-        .listPrincipalTableColumnGrants(principalName, principalType, catName, dbName, tableName, columnName);
-  }
-
-  @Override public List<HiveObjectPrivilege> listPrincipalPartitionColumnGrants(String principalName,
-      PrincipalType principalType, String catName, String dbName, String tableName, List<String> partValues,
-      String partName, String columnName) {
-    return rawStore
-        .listPrincipalPartitionColumnGrants(principalName, principalType, catName, dbName, tableName, partValues,
-            partName, columnName);
-  }
-
-  @Override public boolean grantPrivileges(PrivilegeBag privileges)
-      throws InvalidObjectException, MetaException, NoSuchObjectException {
-    return rawStore.grantPrivileges(privileges);
-  }
-
-  @Override public boolean revokePrivileges(PrivilegeBag privileges, boolean grantOption)
-      throws InvalidObjectException, MetaException, NoSuchObjectException {
-    return rawStore.revokePrivileges(privileges, grantOption);
-  }
-
-  @Override public boolean refreshPrivileges(HiveObjectRef objToRefresh, String authorizer,
-      PrivilegeBag grantPrivileges) throws InvalidObjectException, MetaException, NoSuchObjectException {
-    return rawStore.refreshPrivileges(objToRefresh, authorizer, grantPrivileges);
-  }
-
-  @Override public Role getRole(String roleName) throws NoSuchObjectException {
-    return rawStore.getRole(roleName);
-  }
-
-  @Override public List<String> listRoleNames() {
-    return rawStore.listRoleNames();
-  }
-
-  @Override public List<Role> listRoles(String principalName, PrincipalType principalType) {
-    return rawStore.listRoles(principalName, principalType);
-  }
-
-  @Override public List<RolePrincipalGrant> listRolesWithGrants(String principalName, PrincipalType principalType) {
-    return rawStore.listRolesWithGrants(principalName, principalType);
-  }
-
-  @Override public List<RolePrincipalGrant> listRoleMembers(String roleName) {
-    return rawStore.listRoleMembers(roleName);
-  }
-
   @Override public Partition getPartitionWithAuth(String catName, String dbName, String tblName, List<String> partVals,
       String userName, List<String> groupNames) throws MetaException, NoSuchObjectException, InvalidObjectException {
     catName = StringUtils.normalizeIdentifier(catName);
@@ -2424,67 +2297,6 @@ public class CachedStore implements RawStore, Configurable {
 
   @Override public void setMetaStoreSchemaVersion(String version, String comment) throws MetaException {
     rawStore.setMetaStoreSchemaVersion(version, comment);
-  }
-
-  @Override public List<HiveObjectPrivilege> listPrincipalDBGrantsAll(String principalName,
-      PrincipalType principalType) {
-    return rawStore.listPrincipalDBGrantsAll(principalName, principalType);
-  }
-
-  @Override public List<HiveObjectPrivilege> listPrincipalDCGrantsAll(String principalName,
-                                                                      PrincipalType principalType) {
-    return rawStore.listPrincipalDCGrantsAll(principalName, principalType);
-  }
-
-  @Override public List<HiveObjectPrivilege> listPrincipalTableGrantsAll(String principalName,
-      PrincipalType principalType) {
-    return rawStore.listPrincipalTableGrantsAll(principalName, principalType);
-  }
-
-  @Override public List<HiveObjectPrivilege> listPrincipalPartitionGrantsAll(String principalName,
-      PrincipalType principalType) {
-    return rawStore.listPrincipalPartitionGrantsAll(principalName, principalType);
-  }
-
-  @Override public List<HiveObjectPrivilege> listPrincipalTableColumnGrantsAll(String principalName,
-      PrincipalType principalType) {
-    return rawStore.listPrincipalTableColumnGrantsAll(principalName, principalType);
-  }
-
-  @Override public List<HiveObjectPrivilege> listPrincipalPartitionColumnGrantsAll(String principalName,
-      PrincipalType principalType) {
-    return rawStore.listPrincipalPartitionColumnGrantsAll(principalName, principalType);
-  }
-
-  @Override public List<HiveObjectPrivilege> listGlobalGrantsAll() {
-    return rawStore.listGlobalGrantsAll();
-  }
-
-  @Override public List<HiveObjectPrivilege> listDBGrantsAll(String catName, String dbName) {
-    return rawStore.listDBGrantsAll(catName, dbName);
-  }
-
-  @Override public List<HiveObjectPrivilege> listDCGrantsAll(String dcName) {
-    return rawStore.listDCGrantsAll(dcName);
-  }
-
-  @Override public List<HiveObjectPrivilege> listPartitionColumnGrantsAll(String catName, String dbName,
-      String tableName, String partitionName, String columnName) {
-    return rawStore.listPartitionColumnGrantsAll(catName, dbName, tableName, partitionName, columnName);
-  }
-
-  @Override public List<HiveObjectPrivilege> listTableGrantsAll(String catName, String dbName, String tableName) {
-    return rawStore.listTableGrantsAll(catName, dbName, tableName);
-  }
-
-  @Override public List<HiveObjectPrivilege> listPartitionGrantsAll(String catName, String dbName, String tableName,
-      String partitionName) {
-    return rawStore.listPartitionGrantsAll(catName, dbName, tableName, partitionName);
-  }
-
-  @Override public List<HiveObjectPrivilege> listTableColumnGrantsAll(String catName, String dbName, String tableName,
-      String columnName) {
-    return rawStore.listTableColumnGrantsAll(catName, dbName, tableName, columnName);
   }
 
   @Override public void createFunction(Function func) throws InvalidObjectException, MetaException {
