+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/metastore/metastore/impl/TableStoreImpl.java
@@ -123,10 +123,10 @@ import static org.apache.hadoop.hive.metastore.ObjectStore.convertToParts;
 import static org.apache.hadoop.hive.metastore.ObjectStore.convertToTable;
 import static org.apache.hadoop.hive.metastore.ObjectStore.getJDOFilterStrForPartitionNames;
 import static org.apache.hadoop.hive.metastore.ObjectStore.getPartQueryWithParams;
-import static org.apache.hadoop.hive.metastore.ObjectStore.getPrincipalTypeFromStr;
 import static org.apache.hadoop.hive.metastore.ObjectStore.makeParameterDeclarationString;
 import static org.apache.hadoop.hive.metastore.ObjectStore.putPersistentPrivObjects;
 import static org.apache.hadoop.hive.metastore.ObjectStore.verifyStatsChangeCtx;
+import static org.apache.hadoop.hive.metastore.metastore.impl.PrivilegeStoreImpl.getPrincipalTypeFromStr;
 import static org.apache.hadoop.hive.metastore.utils.MetaStoreUtils.getDefaultCatalog;
 import static org.apache.hadoop.hive.metastore.utils.MetaStoreUtils.newMetaException;
 import static org.apache.hadoop.hive.metastore.utils.StringUtils.normalizeIdentifier;
